fix: don't fetch original intent when disconnected
Prevent crashing in debug builds, when logging in as a new user.
We would be disconnected, but still try to fetch the original intent,
which would fail because that user has not yet connected anywhere.
This resulted in us requesting the tunnel to stop, which would trigger
an assertion failure in debug builds.