fix(ios): dont log in when auth keychain is empty
When determining whether we are logged in on app launch on iOS, we check
the VPNKeychain. We recently introduced a bug where when our session
expired, we cleared the auth keychain but failed to clear the rest. This
commit introduces another check to prevent showing the logged in UI in
this case.

Jira-Id: VPNAPPL-2621