fix(connection): prevent disconnect failures
Ensure that once connection preparation finishes, and `connectionState`
is updated to `.connecting(resolved)`, that the core connection feature
receives the appropriate action to start the tunnel.
Without this, it was possible to start a disconnection request before
the tunnel was started, which cancelled the connection request and we
got stuck in `.connecting`.

Jira-Id: VPNAPPL-2684