fix(nehelper): call read until EOF
readMinimumLength documentation says:

> Read the requested range of bytes. The completion handler will be invoked when:
>	- Exactly "length" number of bytes have been read. 'data' will be non-nil.
>
>	- Fewer than "length" number of bytes, including 0 bytes, have been read, and the connection's
>	read side has been closed. 'data' might be nil, depending on whether there was any data to be
>	read when the connection's read side was closed.
>
>	- Some fatal error has occurred, and 'data' will be nil.
>
>	To know when to schedule a read again, check for the condition whether an error has occurred.

The part that says "the connection's read side has been closed" is
incorrect. There's no way to see if the connection is still open or
whether we should try to read again, so this change just calls
`readMinimumLength` one extra time to see if there is still data.

This most likely addresses many issues with the session expiring
prematurely or the certificate failing to load - in most cases where
there is still more data to be read, all of the headers are sent without
the body, which is available after the second call. This empty response
body would cause errors resulting in the response (certificate, session
token, etc) to be considered invalid.