fix(connection): prevent certificate refresh loop
After receiving an error with the `reconnect.withNewKeysAndCertificate`
`LocalAgentErrorResolutionStrategy`, we get stuck in a loop of clearing
our keys and certificate, asking the network extension for a new
certificate, and again failing to connect because the old private key
in the wireguard configuration no longer matches the new keys we just
generated and used to connect to local agent with.

For now, let's instead remove our keys+certificate from the keychain
and fully disconnect, alerting the user to the error.

Jira-Id: VPNAPPL-2698