fix(connection): remove username/account property
Stop setting the unnecessary username property of
NETunnelProviderProtocol when configuring the tunnel for connection to
a server in the new connection layer. This property is only used for
IKEv2 connection, and is not used when authenticating when connecting
with WG.

This removes the Account field in the VPN configuration visible in iOS
settings.

Jira-Id: VPNAPPL-2742