fix(connection): handle local agent reconnection
There are two ways for the Local Agent client to report that it needs
to be reconnected with a new certificate. We already handled one, where
this was raised through an error event:
(LocalAgentEvent.error(LocalAgentError))

The other way was for the Local Agent to enter an error state:
(LocalAgentEvent.state(LocalAgentState)). Three of these states
represent a situation where we should refresh our certificate (or reload
it, since the network extension should have already refreshed it in the
background). This wasn't being done, and as a result we were not
receiving NetShield stats.

This commit ensures that CoreConnectionFeature now handles this
correctly by reloading the certificate and reconnecting LA.

Jira-Id: VPNAPPL-2759