fix(connection): handle key errors (cert auth)
- Prevent trying to regenerate keys when they are found to be missing
during certificate authentication; this won't do us any good unless we
reconfigure and restart the tunnel (it uses the private key to connect)
- On failing to refresh our certificate (when the extension runs into a
409 conflict error for example), make sure to purge existing keys and
immediately fail the connection attempt instead of timing out after 30
seconds.

The second scenario should not happen after introducing a workaround in
the next commit. The root cause should be solved as part of
VPNAPPL-2766

Jira-Id: VPNAPPL-2757