fix: move server spec filter out of GRDB
Previously we would reconstruct the server filter if a query didn't
return any results in order to figure out which filter step was likely
the problem for a given DB query. This change, while affecting
performance, is meant to give us more accurate errors to display to the
user in the case of a server filter resolution failure.

Jira-Id: VPNAPPL-2677