fix(connection): certificate features
When asking the network extension to refresh our certificate, specify
the current connection features. This ensures that when the extension
reconnects after a connectivity break, and local agent is not running
(e.g. because the app is in the background), the most recently
requested certificate will be retrieved by the vpn server and features
will be correctly applied.

This should make it more likely that features are persistent for the
whole connection duration.

Jira-Id: VPNAPPL-2769