fix(connection): optional ConnectionDetails fields
When DeviceCountry or DeviceIP fields are missing (as is often the case
when connecting to Secure Core servers), don't throw an error, so that
we can still decode and use the exit IP reported by LocalAgent.

This prevents us from showing the potentially inaccurate IP of the
endpoint as stored in the server repository.

Jira-Id: VPNAPPL-2781