fix(connection): hardjail loop (dance party)
Don't immediately disconnect and attempt to reconnect with the same
certificate on transitioning to the clientCertificateError, softJailed
and hardJailed local agent states.

The local agent shared go library follows such state transitions with
an explicit onError delegate call, which contains more specific
information such as the error code. This allows us to handle different,
nuanced scenarios.

Sometimes, we enter one of these states, but refreshing our certificate
is not appropriate (e.g `86108: restrictedServer`). In this case we
should just wait a moment, instead of attempting to reconnect
immediately, which was causing an inifinite loop until we eventually
timed out after 30 seconds.

Jira-Id: VPNAPPL-2800