fix(connection): prevent late ipc retry
Reduce amount of TNAB and UNEX errors by cancelling certificate
refresh effects when disconnecting. When the tunnel is disconnected due
to an unknown reason, we will still fail with an error. But if we abort
the connection attempt due to a timeout, or if the user disconnects
explicitly, we will no longer display an extra error.

Jira-Id: VPNAPPL-2797