fix(macos): too many killswitch changed notifications on appDidBecomeActive
When app was being put in foreground, the dropdown menus are reloaded. This triggers a
property set on the propertiesManager, especially on the killswitch property. Some recent
code addition was listening to updates on this property, even when this was still the same
value. It caused setOnDemand calls on the VpnManager, eventually causing what was appearing
as reconnections.

Jira-Id: VPNAPPL-2953