fix(ios): adjust session timeout to 30s
AsyncTimeout was causing silent logins to fail. In some cases, this
resulted in a timeout happening after the session token was refreshed,
which meant that the subsequent login could _also_ fail since it could
potentially thrash the valid session token.

For now we will increase this timeout to 30s and leave the feature flag
timeout at 5s. We should comprehensively review the login path once
we've successfully moved to Muon.

Jira-Id: VPNAPPL-3200